5
the project is all done. user can now search for top 'n' names in an year, obtain a graph/bar-chart/pie-chart for the popularity of a particular name, inserted by the user along with a particular year (to and from), also inserted by the user. in similar way, user can also compare two names using a ';' in between the two names (eg a;b)